pyproject.toml
[tool.poetry]
name = "ai-image-finder"
version = "0.1.0"
description = ""
authors = ["Alberto Ursino <albertoursino98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ollama = "^0.3.3"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
